successfully implemented the global exception handler

+++ src/main/java/com/laundry/service/CustomerService.java
@@ -2,20 +2,21 @@ package com.laundry.service;
 
 import com.laundry.dto.CustomerRequest;
 import com.laundry.dto.CustomerResponse;
+import com.laundry.exception.BalanceNotSettledException;
+import com.laundry.exception.CustomerNotFoundException;
+import com.laundry.exception.PhoneNumberAlreadyExistException;
 import com.laundry.model.Customer;
 import com.laundry.model.CustomerAccount;
 import com.laundry.repo.CustomerAccountRepository;
 import com.laundry.repo.CustomerRepository;
 import com.laundry.repo.OrderRepository;
 import jakarta.transaction.Transactional;
-import jakarta.validation.constraints.NotBlank;
-import jakarta.validation.constraints.Pattern;
+
 import lombok.RequiredArgsConstructor;
 
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpStatus;
+
 import org.springframework.stereotype.Service;
-import org.springframework.web.server.ResponseStatusException;
+
 
 import java.util.ArrayList;
 import java.util.List;
@@ -34,10 +35,7 @@ public class CustomerService {
     public CustomerResponse addCustomer(CustomerRequest request){
         // Add phone number validation
         if (customerRepo.existsByPhoneNumber(request.getPhoneNumber())) {
-            throw new ResponseStatusException(
-                    HttpStatus.BAD_REQUEST,
-                    "Phone number already exists"
-            );
+            throw new PhoneNumberAlreadyExistException(request.getPhoneNumber());
         }
             Customer customer = new Customer();
             customer.setName(request.getName());
@@ -76,12 +74,12 @@ public class CustomerService {
     public CustomerResponse getCustomerById(Long id) {
         return customerRepo.findById(id)
                 .map(this::mapToResponse)
-                .orElseThrow(() -> new RuntimeException("Customer not found"));
+                .orElseThrow(() -> new CustomerNotFoundException(id));
     }
 
     public CustomerResponse updateCustomer(Long id, CustomerRequest request) {
         Customer customer = customerRepo.findById(id)
-                .orElseThrow(() -> new RuntimeException("Customer not found"));
+                .orElseThrow(() -> new CustomerNotFoundException(id));
 
         customer.setName(request.getName());
         customer.setPhoneNumber(request.getPhoneNumber());
@@ -91,9 +89,13 @@ public class CustomerService {
 
     @Transactional
     public void deleteCustomer(Long id){
-        Customer customer=customerRepo.findById(id).orElseThrow(() -> new RuntimeException("Customer not found"));
+        Customer customer=customerRepo.findById(id).orElseThrow(() -> new CustomerNotFoundException(id));
         System.out.println("Bring it on");
 
+        CustomerAccount account = accountRepo.findByCustomer(customer)
+                .orElseThrow(() -> new CustomerNotFoundException(id));
+
+        if(account.getBalance() !=0.0) throw new BalanceNotSettledException(account.getId(), account.getBalance());
 
         orderRepo.deleteByCustomer(customer);
         System.out.println("Deleted the orders");
